// src/es/schemas/security.suggest_user_profiles.ts
export const SecurityUserProfileUser = z.object({
  email: z.union([z.string(), z.null()]).optional(),
  full_name: z.union([z.lazy(() => Name), z.null()]).optional(),
  realm_name: z.lazy(() => Name),
  realm_domain: z.lazy(() => Name).optional(),
  roles: z.array(z.string()),
  username: z.lazy(() => Username)
}).meta({ id: 'SecurityUserProfileUser' })
export type SecurityUserProfileUser = z.infer<typeof SecurityUserProfileUser>